tree: improve comments to describe raft + prune cases
Originally, this code was developed by going through the cases described in
the paper.  However, a lot of this was lost when the code condensed to what
it is now.  That was a mistake.

For documentation purposes, the cases listed in Wikipedia's description of
Red-black Trees were consulted and comments were added to document the
corresponding cases in this code.  A notable structural difference is that
case 3 in graft and case 3 in prune both follow later cases, looping back to
case 1 on the grandparent for graft and the parent for prune.