tree: add transplant method
This method takes two nodes, one a member and the other a non-member.  Both
nodes should have equal keys, although this requirement can be relaxed such
that the non-member's key must preserve the order of the tree, falling
between the keys for the member's prev and next, inclusive.

After the transplant, the former member will no longer be a member of the
tree and the former non-member will occupy the same location in the tree
structure.

This operation is equivalent to a prune of the member and a graft of the
non-member but is more efficient because the tree structure does not change.