tree: lift color swap out of unlink, through replace, into prune
Previously, unlink would preserve the color of the node.  This is a little
confusing and that semantic is only required in replace, which only required
it as part of prune, which only required it to simplify the code.  Lifting
the color check complicates prune slightly but the readability is much
better and prevents unnecessary toggle-toggles in other places.