test: validate heap
Like order, the values exhumed from heap should ascend.  Therefore, test by
iterating through tree and verifying that the values match.  It is notable
that the special transplant node will not appear in the same place in the
order output as in the heap output.  In the order, it is always the first in
a run of equivalent keys unless it has the maximum key, in which case it is
at the end of the order.  In the heap, it is arbitrarily placed.