new: add intrusive_table
This container implements an intrusive chained hash table, with template
parameters similar to intrusive_tree but with the addition of a hash
function.  When a node is accessed, it is moved to the front of its bucket,
speeding up future access.

Hash buckets are an orthogonal concern and must be provided to the hash
table, either as part of the constructor or during a rehash; the data
structure will not automatically perform a rehash when unbalanced.  If the
lifetime of the provided buckets is shorter than that of the hash table, a
dehash must be performed in order to release them.