program: add dijkstra
This example program consumes a directed graph, performs Dijkstra's
algorithm, and emits paths + costs for every vertex in the graph using heap,
order, & table.

The set of vertices is stored in an intrusive hash table keyed on a string
ID for the vertex.  The table is rehashed when the number of buckets is less
than a load factor times the total number of vertices.  The starting node is
the lexicographically least in the set.

Each vertex contains a set of incoming and outgoing edges, ordered by cost.
The cost of an edge is always non-negative; edges also contain links back to
the vertices at both ends.  Each vertex also contains a route pointer and,
if the route pointer is valid, a total cost through that route.

A heap is used as a priority queue of candidate vertices sorted by cost.
This implementation does not use the DecreaseKey version of the algorithm.

The following data format is used.

	a b 1
	a c 1
	a d 4
	b d 2
	c d 1
	x a 0

The example input produces the following output, sorted after the fact.

	a 0
	b a 1
	c a 1
	d c a 2
	x

In this case, the best path from a to d costs 2 and is via c.  The node x is
not reachable from the starting node.