new: add intrusive_set
This container implements an intrusive disjoint set.  Every node that is a
member of a set is considered typed and is able to return an archetype,
which must be allocated separately.  This operation performs a modified
version of path compression, relinking each node to its grandparent.

Untyped nodes are able to join one and only one existing set and existing
sets may be conjoined, which performs union-by-rank.  In order to release
nodes, an entire set may be dissolved and the former members are free to join
other sets.  To support this operation, the set includes a linked-list of
member nodes; therefore, iteration is also provided.