program: add kruskal
This example program consumes an undirected graph, performs Kruskal's
algorithm, and emits the edges of a minimum spanning tree.  The input data
format is identical to dijkstra.  The program uses heap, set, & table.

The vertices are stored in a table, keyed by a string name, and may be
members of sets.  The edges are stored in a heap, ordered by cost, and
contain a pointer to the vertices at both ends.

For each edge, if both vertices are untyped, a new set is created; if only
one vertex is untyped, it joins the other's archetype set; if the archetypes
of the vertices do not match, they are conjoined; otherwise, skip this edge.