kruskal: refactor unification code into vertex_t
The set unification code is a little tricky, since vertex_t does not create
sets for individuals, and should be a method of the vertex_t class.  This
promotes reusability and significantly simplifies the main loop, making the
algorithm stand out.

In the mysterious future, it is probably worth refactoring the
bucket-management code as well.