heap: add sift method
This method allows one to remove an arbitrary node from the heap.  Unlike
exhume, it is not efficient since it must traverse the entire siblings list
and account for children.  It does allow decrease-key to be implemented as

	heap.sift(n);
	n->key -= x;
	heap.inhume(n);

although there may be more efficient methods, e.g. if heap-order is not
violated.