heap: extract pairing + melding pass
This operation fundamentally takes a node and melds all of its children into
single root.  Currently it is only used for exhume but it could also be used
for removing arbitrary nodes, which might have children.  This refactor has
the added benefit of simplifying the code, which was previously using root
for state unnecessarily.