dijkstra: replace test heap with siftable vertices
Originally, the algorithm was implemented with a priority queue of tests to
perform, since the intrusive heap implementation did not support removal or
priority decrease.  Now it does, so the relax operation can be performed on
the vertices themselves.  This change results in a 20% performance increase
for the main loop, although most time is actually spend on input & output.