dijkstra: output routes, not full paths
Previously, the output of dijkstra was a full path to the source node
followed by the total cost of the route.  This is OK for testing on small
graphs but problematic for processing large graphs; since a worst-case route
might traverse every node, the output is could be quadratic in the number of
nodes.  This change makes it linear instead.

Also: nodes without a route to the source no longer appear in the output.