program: add tarjan
This example program consumes a directed graph, performs a modified version
of Tarjan's algorithm, and emits sets of strongly connected components, one
set per line.  The input format is similar to dijkstra & kruskal, sans an
edge cost.  The program uses set, stack, & table.

The vertices are stored in a table, keyed by a string name, and may be
members of sets.  The edges are stored as an iterable stack inside each
origin vertex.

The biggest modification is that the depth-first search is implemented
iteratively instead of recursively.  This means that each vertex must store
enough state to continue the search.  Unvisited vertices have no index
value.  When initially visited, the vertex's index is set to pre-order
discovery time and the search continues to its edges by recording the
next-searchable egress edge.  These edges are pushed onto the stack one at a
time, ensuring that the search is depth-first.  Once a vertex has no more
edges to search, is is closed.

On close, a vertex's edges, which must have already been opened, are
evaluated in order to discover loops.  Each vertex begins as its own root.
If the other root is not in the stack, then we have only discovered a
cross-edge and move on.  Then, the root pre-order indexes are compared and
if the other root has a smaller pre-order index, this vertex joins that set
of strongly connected components.

Since the set of strongly connected components is stored in a disjoint set,
the running time is not O(V+E) but it is very close.  The set could be
replaced with a stack or a queue to achieve that running time.