queue: add chain overload for partial consumption
This new method is like chain, but only transfers n or fewer queue nodes
from the producer queue to the consumer queue.  This is equivalent to doing

	for (int i = 0 ; !producer.empty() && i < n ; ++i)
		consumer.enqueue(producer.dequeue());

but avoids all of the superfluous accounting.  Since the existing chain
method essentially does all of this for infinite n, it feels like a logical
extension and is useful for implementing algorithms like merge sort.