queue: add sort method template
This method template implements merge sort for the items in a queue, given
some member pointer and a compare function, with a default of compare<K>.
Many existing merge-sort-for-list implementations are recursive, but could
easily be done iteratively, as here.

The basic algorithm is to pull a certain number of queue nodes, doubling the
size each iteration, and to merge them into an accumulator queue, creating a
sequence of sorted nodes double the original size.  Once the pull size is
greater than the total size, the queue is sorted.

This implementation infers that by pulling queue nodes and then checking to
see whether both the remaining to-sort queue and the accumulator are empty;
this could be improved by calculating the total queue size on the first
iteration and stopping preëmptively when the pull size is larger.