queue: factor out merge method template
Previously, this was implemented in sort directly.  However, it would be
useful to merge known-sorted queues together.  Rather than do this in place,
the nodes of each sorted queue are chained to a destination queue, on which
there are no requirements.  It would be trivial to overload these templates
to support that behavior and I wanted to minimize the impact on the existing
code.

I am considering bundling all of this up into a utility sub-class very soon
in anticipation of refactoring intrusive_order.