queue: refactor sort + friends into utility sub-class
Yes, I've heard that utility classes are evil.  In this case, it makes sense
here for a few reasons.

First, if the compiler does not implement C++11 and therefore does not
support default parameters for function templates, it avoids helper template
functions that specify the default comparison function for sort, merge, &
sorted.

Second, it prevents repetition of the template parameters via a typedef in
the common case where a single key and comparison function never change.

Third, the algorithm implementation is very specific to this queue template
and does not generalize for other containers.