order: remove silly class
This class is not that great.  It was originaly developed because I used to
ask an interview question re: linked lists and sorting where this was one of
the obvious solutions despite the quadratic complexity.

Since queue::sorted has been added, it is a much better replacement.  I
wanted to reimplement order in terms of a sorted and unsorted queue that
were merged at an opportune time.  Unfortunately, C++ does not support
generating pointer-to-members in terms of other pointer-to-mermbers as
compile time constants.  This means, that an order link could not contain
both queue links and then instantiate queues from them based on the order
link passed in the template parameter.  P0149R0 or a similar proposal might
address this but is not yet part of the standard.