heap: add rehume method
This method checks whether a node is violating the heap-order property wrt
its parent and, if so, resolves the issue.  This allows key-decrease to to
be written as

	n->key -= x;
	heap.rehume(n);

instead of

	heap.sift(n);
	n->key -= x;
	heap.inhume(n);

and while it does mean that the sibling chain must be followed to find the
parent, it avoids an optimization pass of any children and the expensive
sift+inhume in the case that the heap-order is not violated.

This method does not work in the case of a key-increase, since the entire
sub-tree would need to be checked and that is in general more expensive than
just doing the sift+inhume.