tree: add fell method
This method takes a method pointer and traverses the tree, unlinking each
node and then invoking the method.  This makes it easy to destroy a tree
without having to prune and rebalance.  It is similar to table::wipe and
set::disolve.  Like those methods, the axe is responsible for cleaning up
the unlinked node.

Because of the way unlink works, each subtree must be destroyed prior to the
parent node.  Therefore, the iteration is identical to inosculate.