heap: add bury method
Whereas rehume allows one to do a key-decrease, bury allows one to do a
key-increase without having to sift+inhume.  If the node has any children,
they are run through a pairing pass and the resulting node is made a sibling
of the former parent.  This avoids a full sibling walk.