heap: better bury balance
In heap::bury, the children were always being passed and the resulting root
was converted to a sibling.  In the case where this root is larger than the
target node, it can actually remain the sole child if the heap property is
not violated.  Furthermore, immediately subsequent bury operations will
either maintain the status quo OR convert the single child to a single
sibling but the pass will become trivial.

This method could also be special-cased to avoid the unlink+relink pair in
the case of an only child and no violation, but the obvious solution starts
to duplicate code.

In order to stress this case, bury with no value change.