tree: fix next + prev
The loop condition for a right-side ascension was incorrect in next.  In
particular, if we were iterating from the root but the right subtree was
empty then we would ask whether the root was the left node of its parent,
which is meaningless and therefore asserts.  The loop condition would
prevent this from happening on any other iteration, since it would bail if
the current node was root.

Instead, attempt to ascend if there is a parent and check the sided-ness
directly.  If we ascend to the root, then we are coming from the rightest
node and the iteration is complete.

This bug exists symmetrically for prev and is also addressed by this commit.